swe_worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyswisseph==2.10.3.2
